% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{nr_add_node_path}
\alias{nr_add_node_path}
\title{Add the path to node/npm to the system path}
\usage{
nr_add_node_path()
}
\value{
\code{TRUE} on success, invisibly
}
\description{
If \code{npm} can already be seen by the system, the system path will not be changed. Otherwise, the system path will be prepended with the node/npm path.
}
