% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{nr_install_node}
\alias{nr_install_node}
\title{Install nodeJS}
\usage{
nr_install_node(version = "LTS", bits, force = FALSE)
}
\arguments{
\item{version}{string: "LTS" (long-term support, recommended) or "current"}

\item{bits}{integer: 32 or 64, for 32- or 64-bit install. If missing or \code{NULL}, will be guessed based on \code{.Machine$sizeof.pointer}}

\item{force}{logical: force reinstallation if node already exists}
}
\value{
the path to the installed executable
}
\description{
This is a helper function to install node. The node binaries will be downloaded from \url{https://nodejs.org/en/download/} and saved to your user appdata directory.
}
\examples{
\dontrun{
  nr_install_node()
}

}
\references{
\url{https://nodejs.org/en/download/}
}
