% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{nr_node_exe}
\alias{nr_node_exe}
\title{The path to the node executable}
\usage{
nr_node_exe()
}
\value{
The path to the executable, or \code{NULL} if not found
}
\description{
The path to the node executable
}
\examples{

nr_node_exe()

}
